myapp/requirements.txt
Flask
pytesseract
Pillow
tensorflow